requirements.txt
aiohttp==3.6.2  # 3.6.2+ fixes bug with unpickleable exceptions
numpy>=1.17.0
comet_ml
progress
Pillow
fs
glob2
sklearn
plotly
imageio
imageio-ffmpeg
pytest
aiobotocore
botocore
boto3
hanging_threads
matplotlib
seaborn
opencv-python-headless
opencv-contrib-python
psutil
requests>=2.22.0
deprecated
tensorflow-probability==0.7
setproctitle
fs_s3fs_ng
absl-py
tenacity
jupyter
s4cmd
smart_open>=5.1.0
recordtype
pyyaml
shyaml
boto3-stubs
colour
uvloop
awswrangler
humanize
diskcache
statsmodels
plotting
chart-studio
# Since adding Lmer model
tensorflow-gpu
